+++ dev/examples/manifold/fdg.clj
@@ -63,6 +63,7 @@
 ;; Parametrized torus:
 
 (p/scene
+ {:threestrap {:plugins  ["core" "controls" "cursor" "stats"]}}
  (ev/with-let [!opts {:R Math/PI :r Math/PI}]
    [:<>
     (leva/controls