+++ dev/examples/mathbox/functions.clj
@@ -22,6 +22,7 @@
 ;;
 ;; Let's make a function to try!
 
+
 (defn my-fn [x]
   (+ -1
      (square (sin x))